requirements.txt
Flask
scikit-learn
numpy
pandas
pdfminer.six
python-docx
PyPDF2
Pillow
imagehash
opencv-python
scikit-image
tqdm
gunicorn
